# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * DelhiGhar API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Listing {
  id: string;
  name: string;
  type: string;
  gender: string;
  area: string;
  gali: string;
  landmark: string;
  /** @nullable */
  mapsLink?: string | null;
  rent: number;
  deposit: number;
  roomType: string;
  ownerName: string;
  ownerPhone: string;
  /** @nullable */
  alternatePhone?: string | null;
  verificationStatus: string;
  /** @nullable */
  aadhaarImageUrl?: string | null;
  escrowEnabled: boolean;
  amenities: string[];
  securityFeatures: string[];
  curfewTime: string;
  guestPolicy: string;
  smokingAllowed: boolean;
  alcoholAllowed: boolean;
  nonVegAllowed: boolean;
  photos: string[];
  visitCount: number;
  fraudReportCount: number;
  rating: number;
  reviewCount: number;
  isHidden: boolean;
  createdAt: string;
}

export interface ListingInput {
  name: string;
  type: string;
  gender: string;
  area: string;
  gali: string;
  landmark: string;
  mapsLink?: string;
  rent: number;
  deposit: number;
  roomType: string;
  ownerName: string;
  ownerPhone: string;
  alternatePhone?: string;
  aadhaarConsent?: boolean;
  aadhaarImageUrl?: string;
  escrowEnabled: boolean;
  amenities: string[];
  securityFeatures: string[];
  curfewTime: string;
  guestPolicy: string;
  smokingAllowed: boolean;
  alcoholAllowed: boolean;
  nonVegAllowed: boolean;
  photos: string[];
}

export interface ListingUpdate {
  name?: string;
  rent?: number;
  deposit?: number;
  verificationStatus?: string;
  isHidden?: boolean;
  fraudReportCount?: number;
}

export interface ListingStats {
  totalListings: number;
  verifiedListings: number;
  pgCount: number;
  flatCount: number;
  hostelCount: number;
  avgRent: number;
  escrowEnabledCount: number;
  recentListings: Listing[];
}

export interface SaveInput {
  userId: string;
  action: string;
}

export interface SaveResult {
  saved: boolean;
}

export interface RevealPhoneInput {
  userId: string;
}

export interface PhoneRevealResult {
  phone: string;
  /** @nullable */
  alternatePhone?: string | null;
}

export interface VisitInput {
  listingId: string;
  buyerName: string;
  buyerPhone: string;
  visitDate: string;
  userId: string;
}

export interface Visit {
  id: string;
  listingId: string;
  buyerName: string;
  buyerPhone: string;
  visitDate: string;
  status: string;
  createdAt: string;
}

export interface EscrowOrderInput {
  listingId: string;
  buyerId: string;
  amount: number;
}

export interface EscrowOrder {
  orderId: string;
  amount: number;
  currency: string;
  keyId: string;
}

export interface PaymentConfirmInput {
  orderId: string;
  paymentId: string;
  signature: string;
  listingId: string;
  buyerId: string;
}

export interface EscrowActionInput {
  paymentId: string;
  listingId: string;
  password?: string;
}

export interface PaymentResult {
  success: boolean;
  message: string;
}

export interface ReportInput {
  listingId: string;
  reporterId: string;
  reasons: string[];
  details?: string;
}

export interface Report {
  id: string;
  listingId: string;
  reporterId: string;
  reasons: string[];
  /** @nullable */
  details?: string | null;
  createdAt: string;
}

export interface AadhaarVerifyInput {
  action: string;
  password: string;
}

export type GetListingsParams = {
type?: string;
gender?: string;
budgetMin?: number;
budgetMax?: number;
verifiedOnly?: string;
search?: string;
};

export type GetSavedListingsParams = {
userId: string;
};

export type GetAdminListingsParams = {
password: string;
status?: string;
};