# lib/api-zod/src/generated/types/escrowActionInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * DelhiGhar API specification
 * OpenAPI spec version: 0.1.0
 */

export interface EscrowActionInput {
  paymentId: string;
  listingId: string;
  password?: string;
}