# lib/api-zod/src/generated/types/escrowOrder.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * DelhiGhar API specification
 * OpenAPI spec version: 0.1.0
 */

export interface EscrowOrder {
  orderId: string;
  amount: number;
  currency: string;
  keyId: string;
}