# lib/api-zod/src/generated/types/listing.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * DelhiGhar API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Listing {
  id: string;
  name: string;
  type: string;
  gender: string;
  area: string;
  gali: string;
  landmark: string;
  /** @nullable */
  mapsLink?: string | null;
  rent: number;
  deposit: number;
  roomType: string;
  ownerName: string;
  ownerPhone: string;
  /** @nullable */
  alternatePhone?: string | null;
  verificationStatus: string;
  /** @nullable */
  aadhaarImageUrl?: string | null;
  escrowEnabled: boolean;
  amenities: string[];
  securityFeatures: string[];
  curfewTime: string;
  guestPolicy: string;
  smokingAllowed: boolean;
  alcoholAllowed: boolean;
  nonVegAllowed: boolean;
  photos: string[];
  visitCount: number;
  fraudReportCount: number;
  rating: number;
  reviewCount: number;
  isHidden: boolean;
  createdAt: string;
}