# lib/api-zod/src/generated/types/listingInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * DelhiGhar API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ListingInput {
  name: string;
  type: string;
  gender: string;
  area: string;
  gali: string;
  landmark: string;
  mapsLink?: string;
  rent: number;
  deposit: number;
  roomType: string;
  ownerName: string;
  ownerPhone: string;
  alternatePhone?: string;
  aadhaarConsent?: boolean;
  aadhaarImageUrl?: string;
  escrowEnabled: boolean;
  amenities: string[];
  securityFeatures: string[];
  curfewTime: string;
  guestPolicy: string;
  smokingAllowed: boolean;
  alcoholAllowed: boolean;
  nonVegAllowed: boolean;
  photos: string[];
}