# lib/api-zod/src/generated/types/listingUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * DelhiGhar API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ListingUpdate {
  name?: string;
  rent?: number;
  deposit?: number;
  verificationStatus?: string;
  isHidden?: boolean;
  fraudReportCount?: number;
}