# lib/api-zod/src/generated/types/saveInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * DelhiGhar API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SaveInput {
  userId: string;
  action: string;
}